+++ mandown/comic.py
@@ -1,62 +1,7 @@
-from dataclasses import dataclass, field
-from pathlib import Path
+from dataclasses import dataclass
 
-from slugify import slugify
-
-from . import iohandler, sources
-
-
-@dataclass(frozen=True)
-class BaseMetadata:
-    title: str
-    authors: list[str]
-    url: str
-    genres: list[str]
-    description: str
-    cover_art: str
-
-    def __post_init__(self) -> None:
-        self.title_slug = slugify(self.title)
-
-    def asdict(self) -> dict:
-        return {
-            "title": self.title,
-            "authors": self.authors,
-            "url": self.url,
-            "genres": self.genres,
-            "description": self.description,
-            "cover_art": self.cover_art,
-        }
-
-
-@dataclass
-class BaseChapter:
-    title: str
-    url: str
-    slug: str = None
-
-    def __post_init__(self) -> None:
-        if self.slug is None:
-            self.slug = slugify(self.title)
-
-    def asdict(self) -> dict:
-        return {
-            "title": self.title,
-            "url": self.url,
-            "slug": self.slug,
-        }
-
-
-@dataclass
-class ComicChapter(BaseChapter):
-    _images: list[str] = field(init=False, default_factory=list)
-
-    @property
-    def images(self) -> list[str]:
-        if self._images:
-            return self.images
-
-        self._images = 0
+from . import sources
+from .base import BaseChapter, BaseMetadata
 
 
 @dataclass
@@ -78,6 +23,3 @@ class Comic:
 
     def get_chapter_image_urls(self, chapter: BaseChapter) -> list[str]:
         return self.source.fetch_chapter_image_list(chapter)
-
-    def save_metadata(self, path: Path | str) -> None:
-        iohandler.save_comic(self, path)
