+++ mandown/sources/source_mangasee.py
@@ -10,7 +10,7 @@ import feedparser
 import requests
 from bs4 import BeautifulSoup
 
-from ..comic import BaseChapter, BaseMetadata
+from ..base import BaseChapter, BaseMetadata
 from .base_source import BaseSource
