+++ mandown/sources/source_webtoons.py
@@ -9,7 +9,7 @@ import feedparser
 import requests
 from bs4 import BeautifulSoup
 
-from ..comic import BaseChapter, BaseMetadata
+from ..base import BaseChapter, BaseMetadata
 from .base_source import BaseSource
